fix(columine): post-parity quirk sweep — every preserved Zig quirk fixed on its merits
Correctness: full u64 aggregate/count rollback (prev_value low + key high
journal lanes; deleted Zig truncated to u32); scalar writes journaled
(ScalarUpdate op 14, a named post-parity ABI extension in the opcode
audit); rollforward skips zeroed no-op redo markers (undo-only lanes)
instead of panicking; reset zeroes the buffer (fresh-allocation
contract); intern buffers grow on demand with stable handles;
struct_map find_insert and nested resolve unified on probe-then-reuse
(no shadowed duplicates, resolve/lookup symmetric); nested regrowth
sets INSERTED only on genuine inserts.

API honesty: exactly-capacity batches accepted (JSON + msgpack, both
arms); $extra with an unusable work buffer refuses loudly; ep-wasm
honors requested capacity (Zig silently clamped to 256 events);
ARROW_PASSTHROUGH refuses uniformly; DynamicColumn::int64 tags Int64.

Unifications: ONE timestamp parser (strict digits, digit-only fractions
reject, first three digits are the ms, 1970..=2099); scalar predicated
min/max on the NaN-skipping kernel policy; one authoritative layout
helper set (unpadded types.zig twin deleted). Kept deliberately,
now WHY-documented maintenance algebra (sum/count group-invertible,
min/max support-scan class): tie-returns-second-operand FP profile,
zeroed nested AGG inners with count==0 emptiness.

+++ packages/columine/crates/columine-arrow/src/record_batch.rs
@@ -156,11 +156,10 @@ impl<'a> DynamicColumn<'a> {
         }
     }
 
-    /// ZIG-PARITY: dynamic_record_batch.zig's `int64` constructor tags the
-    /// column `.Int32` (zig line 164). Harmless — both are "fixed-width,
-    /// data buffer only" in `add_column`'s switch — but the tag is kept
-    /// verbatim so any future per-type branching reproduces Zig; intended
-    /// fix at the post-parity sweep: tag Int64.
+    /// Tags the column `.Int64` (the deleted Zig tagged `.Int32` — harmless
+    /// while `add_column` treats both as "fixed-width, data buffer only",
+    /// but a lie any future per-type branching would inherit; fixed
+    /// post-parity).
     pub fn int64(
         field_idx: u32,
         nullable: bool,
@@ -169,7 +168,7 @@ impl<'a> DynamicColumn<'a> {
     ) -> Self {
         Self {
             field_idx,
-            arrow_type: ArrowType::Int32,
+            arrow_type: ArrowType::Int64,
             nullable,
             validity,
             data,
@@ -519,8 +518,8 @@ mod tests {
         assert_eq!(col.arrow_type, ArrowType::Utf8);
         assert!(!col.nullable);
         let int_col = DynamicColumn::int64(2, false, None, &[]);
-        // ZIG-PARITY tag (see constructor doc).
-        assert_eq!(int_col.arrow_type, ArrowType::Int32);
+        // Post-parity fix pin: the tag tells the truth (Zig tagged Int32).
+        assert_eq!(int_col.arrow_type, ArrowType::Int64);
         assert!(int_col.offsets.is_none());
     }
